순수 함수를 별도 utils 파일로 분리하여 CI 테스트 실패 수정
Firebase 초기화 없이 테스트 가능하도록 buildPostCardDataMap,
deduplicateAuthorIds를 batchPostCardDataUtils.ts로 이동
